Task -1 & Task -2 of Lab -1 had been completed
Lab - 1:
Taks - 1: When someone Run the project it will be shown my name in the emulator & in the logcat my name also be shown

Task -2 : Here a design of the emulator of with buttonview & textview. The design has both horizintal & vertical view in the emulator.  When the emulator rotate the design will shown as where the device is horizontally or vertically.